services/backend/requirements.txt
aerich==0.7.1
asyncpg==0.27.0
bcrypt==4.0.1
email-validator==1.3.1
fastapi==0.95.0
fastapi-mail==1.2.8
filetype==1.2.0
flake8==6.0.0
passlib==1.7.4
Pillow==9.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
tortoise-orm==0.19.3
uvicorn==0.21.1
